+++ vimiv/config/keyfile.py
@@ -7,11 +7,13 @@
 """Methods to read keybindings from file and store them."""
 
 import configparser
-import logging
 import os
 
 from vimiv import api
-from vimiv.utils import xdg
+from vimiv.utils import xdg, log
+
+
+_logger = log.module_logger(__name__)
 
 
 def parse(args):
@@ -35,6 +37,7 @@ def parse(args):
 
 def dump():
     """Write default keybindings to keys file."""
+    _logger.debug("Dumping default keybindings to file")
     parser = KeyfileParser(delimiters=":")
     # Add sections
     parser.add_section("GLOBAL")
@@ -51,7 +54,7 @@ def dump():
     user_file = xdg.join_vimiv_config("keys.conf")
     with open(user_file, "w") as f:
         parser.write(f)
-    logging.debug("Created default keys file %s", user_file)
+    _logger.debug("Created default keys file '%s'", user_file)
 
 
 def _read(filename):
@@ -60,6 +63,7 @@ def _read(filename):
     Args:
         filename: Name of the keybinding file to read.
     """
+    _logger.debug("Reading keybindings from '%s'", filename)
     parser = KeyfileParser()
     parser.read(filename)
     for mode, bindings in api.keybindings.items():
@@ -67,8 +71,8 @@ def _read(filename):
             section = parser[mode.name.upper()]
             _update_bindings(bindings, section)
         except KeyError:
-            logging.info("Missing section '%s' in keys.conf", mode.name.upper())
-    logging.debug("Read keybindings from '%s'", filename)
+            _logger.debug("Missing section '%s' in keys.conf", mode.name.upper())
+    _logger.debug("Read keybindings from '%s'", filename)
 
 
 class KeyfileParser(configparser.ConfigParser):
@@ -94,3 +98,9 @@ def _update_bindings(bindings, section):
     """
     for keybinding, command in section.items():
         bindings[keybinding] = command
+        _logger.debug(
+            "Read keybinding '%s': '%s' for mode '%s'",
+            keybinding,
+            command,
+            section.name,
+        )