+++ vimiv/startup.py
@@ -11,8 +11,6 @@ Module Attributes:
         object must exist until vimiv exits.
 """
 
-import logging
-import logging.handlers
 import os
 import sys
 import tempfile
@@ -23,13 +21,14 @@ from vimiv import app, api, parser, imutils, plugins, version, gui
 from vimiv.commands import runners
 from vimiv.completion import completionmodels
 from vimiv.config import configfile, keyfile, styles
-from vimiv.utils import xdg, crash_handler, statusbar_loghandler, trash_manager
+from vimiv.utils import xdg, crash_handler, log, trash_manager
 
 # Must be imported to create the commands using the decorators
 from vimiv.commands import misccommands  # pylint: disable=unused-import
 
 
 _tmpdir = None
+_logger = log.module_logger(__name__)
 
 
 def main():
@@ -38,10 +37,10 @@ def main():
     qapp = app.Application()
     crash_handler.CrashHandler(qapp)
     setup_post_app(args)
-    logging.debug("Startup completed, starting Qt main loop")
+    _logger.debug("Startup completed, starting Qt main loop")
     returncode = qapp.exec_()
     plugins.cleanup()
-    logging.debug("Exiting with status %d", returncode)
+    _logger.debug("Exiting with status %d", returncode)
     return returncode
 
 
@@ -59,10 +58,10 @@ def setup_pre_app(argv):
         print(version.info())
         sys.exit(0)
     init_directories(args)
-    setup_logging(args.log_level)
-    logging.debug("Start: vimiv %s", " ".join(argv))
-    logging.debug("%s\n", version.info())
-    logging.debug("%s\n", version.paths())
+    log.setup_logging(args.log_level, *args.debug)
+    _logger.debug("Start: vimiv %s", " ".join(argv))
+    _logger.debug("%s\n", version.info())
+    _logger.debug("%s\n", version.paths())
     update_settings(args)
     trash_manager.init()
     return args
@@ -79,35 +78,6 @@ def setup_post_app(args):
     init_paths(args)
     if args.command:
         run_startup_commands(*args.command)
-
-
-def setup_logging(log_level):
-    """Prepare the python logging module.
-
-    Sets it up to write to stderr and $XDG_DATA_HOME/vimiv/vimiv.log.
-
-    Args:
-        log_level: Log level as string as given from the command line.
-    """
-    log_format = logging.Formatter(
-        "[%(asctime)s] %(levelname)s: %(message)s", datefmt="%H:%M:%S"
-    )
-
-    logger = logging.getLogger()
-    logger.handlers = []
-    logger.setLevel(logging.DEBUG)
-
-    file_handler = logging.FileHandler(xdg.join_vimiv_data("vimiv.log"), mode="w")
-    file_handler.setFormatter(log_format)
-    logger.addHandler(file_handler)
-
-    console_handler = logging.StreamHandler()
-    console_handler.setFormatter(log_format)
-    console_handler.setLevel(log_level)
-    logger.addHandler(console_handler)
-
-    statusbar_loghandler.setLevel(log_level)
-    logger.addHandler(statusbar_loghandler)
 
 
 def init_directories(args):
@@ -137,11 +107,11 @@ def init_directories(args):
 
 def init_paths(args):
     """Open paths given from commandline or fallback to library if set."""
-    logging.debug("Opening paths")
+    _logger.debug("Opening paths")
     try:
         api.open(os.path.abspath(os.path.expanduser(p)) for p in args.paths)
     except api.commands.CommandError:
-        logging.debug("init_paths: No valid paths retrieved")
+        _logger.debug("init_paths: No valid paths retrieved")
         if api.settings.startup_library.value:
             api.open([os.getcwd()])
     api.status.update()
@@ -149,7 +119,7 @@ def init_paths(args):
 
 def init_ui(args):
     """Initialize the Qt UI."""
-    logging.debug("Initializing UI")
+    _logger.debug("Initializing UI")
     mw = gui.MainWindow()
     if args.fullscreen:
         mw.fullscreen()
@@ -180,9 +150,9 @@ def update_settings(args):
             setting = api.settings.get(option)
             setting.value = value
         except KeyError:
-            logging.error("Unknown setting %s", option)
+            log.error("Unknown setting %s", option)
         except ValueError as e:
-            logging.error(str(e))
+            log.error(str(e))
 
 
 def run_startup_commands(*commands: str):
@@ -193,9 +163,9 @@ def run_startup_commands(*commands: str):
     """
     total = len(commands)
     for i, command in enumerate(commands):
-        logging.debug("Startup commands: running %d/%d '%s'", i + 1, total, command)
+        _logger.debug("Startup commands: running %d/%d '%s'", i + 1, total, command)
         if "quit" in command:  # This does not work without a running app
-            logging.warning("Quitting forcefully as the app does not exist")
+            log.warning("Quitting forcefully as the app does not exist")
             sys.exit(2)
         else:
             runners.run(command, mode=api.modes.current())